// packages/solid-webgpu-gltf/src/generated/glTF.ts
export type Accessor = GlTFChildOfRootProperty & {
  /**
   * The index of the bufferView.
   */
  bufferView?: GlTFId
  /**
   * The offset relative to the start of the buffer view in bytes.
   */
  byteOffset?: number
  /**
   * The datatype of the accessor's components.
   */
  componentType: 5120 | 5121 | 5122 | 5123 | 5125 | 5126 | number
  /**
   * Specifies whether integer data values are normalized before usage.
   */
  normalized?: boolean
  /**
   * The number of elements referenced by this accessor.
   */
  count: number
  /**
   * Specifies if the accessor's elements are scalars, vectors, or matrices.
   */
  type: 'SCALAR' | 'VEC2' | 'VEC3' | 'VEC4' | 'MAT2' | 'MAT3' | 'MAT4' | string
  /**
   * Maximum value of each component in this accessor.
   *
   * @minItems 1
   * @maxItems 16
   */
  max?: number[]
  /**
   * Minimum value of each component in this accessor.
   *
   * @minItems 1
   * @maxItems 16
   */
  min?: number[]
  /**
   * Sparse storage of elements that deviate from their initialization value.
   */
  sparse?: AccessorSparse
  name?: unknown
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
export type GlTFChildOfRootProperty = GlTFProperty & {
  /**
   * The user-defined name of this object.
   */
  name?: string
  [k: string]: unknown
}
export type GlTFId = number
/**
 * Sparse storage of accessor values that deviate from their initialization value.
 */
export type AccessorSparse = GlTFProperty & {
  /**
   * Number of deviating accessor values stored in the sparse array.
   */
  count: number
  /**
   * An object pointing to a buffer view containing the indices of deviating accessor values. The number of indices is equal to `count`. Indices **MUST** strictly increase.
   */
  indices: AccessorSparseIndices
  /**
   * An object pointing to a buffer view containing the deviating accessor values.
   */
  values: AccessorSparseValues
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * An object pointing to a buffer view containing the indices of deviating accessor values. The number of indices is equal to `accessor.sparse.count`. Indices **MUST** strictly increase.
 */
export type AccessorSparseIndices = GlTFProperty & {
  /**
   * The index of the buffer view with sparse indices. The referenced buffer view **MUST NOT** have its `target` or `byteStride` properties defined. The buffer view and the optional `byteOffset` **MUST** be aligned to the `componentType` byte length.
   */
  bufferView: GlTFId
  /**
   * The offset relative to the start of the buffer view in bytes.
   */
  byteOffset?: number
  /**
   * The indices data type.
   */
  componentType: 5121 | 5123 | 5125 | number
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * An object pointing to a buffer view containing the deviating accessor values. The number of elements is equal to `accessor.sparse.count` times number of components. The elements have the same component type as the base accessor. The elements are tightly packed. Data **MUST** be aligned following the same rules as the base accessor.
 */
export type AccessorSparseValues = GlTFProperty & {
  /**
   * The index of the bufferView with sparse values. The referenced buffer view **MUST NOT** have its `target` or `byteStride` properties defined.
   */
  bufferView: GlTFId
  /**
   * The offset relative to the start of the bufferView in bytes.
   */
  byteOffset?: number
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
// ...
export interface GlTFProperty {
  extensions?: Extension
  extras?: Extras
  [k: string]: unknown
}